# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    Archetype:
      type: string
      enum:
      - EPIC
      - TICKET
      - FEATURE
      - TASK
      - CAMPAIGN
...
    CampaignCriterionProgressDto:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        seeded:
          type: boolean
        satisfied:
          type: boolean
        wouldBeSatisfiedBy:
          type: string
        satisfiable:
          type: boolean
        reason:
          type: string
        evidence:
          $ref: "#/components/schemas/CriterionEvidenceDto"
        approval:
          $ref: "#/components/schemas/CriterionApprovalDto"
        satisfiedAt:
          $ref: "#/components/schemas/Instant"
    CampaignDispatchDto:
      type: object
      properties:
        workspaceId:
          type: string
        branch:
          type: string
        agentLaunch:
          type: string
...
    CampaignEvaluatorDto:
      type: object
      properties:
        connected:
          type: boolean
        lastSweepCompletedAt:
          $ref: "#/components/schemas/Instant"
        stalled:
          type: boolean
        consumerFailing:
          type: boolean
        lastError:
          type: string
        lastErrorAt:
          $ref: "#/components/schemas/Instant"
        watermarkAt:
          $ref: "#/components/schemas/Instant"
    CampaignGroupProgressDto:
      type: object
      properties:
        id:
          type: string
        satisfied:
          type: boolean
        criteria:
          type: array
          items:
            $ref: "#/components/schemas/CampaignCriterionProgressDto"
...
    CampaignMemberEntityDto:
      type: object
      properties:
        id:
          type: string
        archetype:
          type: string
        qualifiedId:
          type: string
        title:
          type: string
        status:
          type: string
        blocked:
          type: boolean
    CampaignMemberProgressDto:
      type: object
      properties:
        membershipId:
          type: string
        position:
          type: integer
          format: int32
        entity:
          $ref: "#/components/schemas/CampaignMemberEntityDto"
        state:
          $ref: "#/components/schemas/CampaignMemberState"
        waitsFor:
          type: array
          items:
            type: string
        joinedRunning:
          type: boolean
        groups:
          type: array
          items:
            $ref: "#/components/schemas/CampaignGroupProgressDto"
        dispatchedAt:
          $ref: "#/components/schemas/Instant"
        dispatch:
          $ref: "#/components/schemas/CampaignDispatchDto"
        dispatchRefusal:
          type: string
        dispatchRefusedAt:
          $ref: "#/components/schemas/Instant"
        dispatchError:
          type: string
...
    CampaignMemberState:
      type: string
      enum:
      - DROPPED
      - DONE
      - DISPATCH_FAILED
      - JOINED_RUNNING
      - RUNNING
      - REFUSED
      - READY
      - WAITING
    CampaignProgressCampaignDto:
      type: object
      properties:
        id:
          type: string
        qualifiedId:
          type: string
        title:
          type: string
        status:
          type: string
        blocked:
          type: boolean
        start:
          $ref: "#/components/schemas/CampaignStartDto"
    CampaignProgressDto:
      type: object
      properties:
        campaign:
          $ref: "#/components/schemas/CampaignProgressCampaignDto"
        evaluator:
          $ref: "#/components/schemas/CampaignEvaluatorDto"
        members:
          type: array
          items:
            $ref: "#/components/schemas/CampaignMemberProgressDto"
    CampaignProgressResponse:
      type: object
      properties:
        progress:
          $ref: "#/components/schemas/CampaignProgressDto"
...
    CampaignStartDto:
      type: object
      properties:
        firstStartedAt:
          $ref: "#/components/schemas/Instant"
        startedAt:
          $ref: "#/components/schemas/Instant"
        startedBy:
          type: string
        active:
          type: boolean
...
    CriterionApprovalDto:
      type: object
      properties:
        approvedBy:
          type: string
        note:
          type: string
...
    CriterionEvidenceDto:
      type: object
      properties:
        eventId:
          type: string
        signature:
          type: string
        summary:
          type: string
...
    DispatchMode:
      type: string
      enum:
      - FLOW
      - PHASE
...
    DispatchRequest:
      type: object
      properties:
        mode:
          type: string
...
    EntityCreate:
      type: object
      required:
      - archetype
      description: "A new entity of any archetype: the archetype, plus exactly the\
        \ body GET /projects/api/entities/archetypes/{archetype}/schemas/create describes\
        \ for it. A root (EPIC, TICKET, CAMPAIGN) names its project, a node (FEATURE,\
        \ TASK) its parent."
      properties:
        archetype:
          $ref: "#/components/schemas/Archetype"
          type: string
          description: "EPIC, TICKET, FEATURE, TASK or CAMPAIGN"
        project:
          type: string
          description: "A root's project: its id or its slug."
        parent:
          type: string
          description: "A node's parent — an EPIC for a FEATURE, a FEATURE for a TASK\
            \ — by UUID or qualified id."
        title:
          type: string
          description: The label.
        description:
          type: string
          description: The long-form Markdown body.
        ticketType:
          type: string
          description: "A ticket's kind, BUG or IMPROVEMENT. MAINTENANCE is reserved\
            \ for the tickets the platform files about its own stuck release requests,\
            \ and it closes those itself."
        impetus:
          type: string
          description: Why a ticket came about. Required for a ticket.
        assignee:
          type: string
          description: An epic's or a ticket's assignee.
        repositoryId:
          type: string
          description: "A task's repository, in the task's project."
        dependsOn:
          type: string
          description: A feature's or task's sibling dependency.
    EntityDispatchDto:
      type: object
      properties:
        entityId:
          type: string
        archetype:
          type: string
        phase:
          type: string
        mode:
          $ref: "#/components/schemas/DispatchMode"
        workspaceRowId:
          type: integer
          format: int64
        repositoryId:
          type: string
        branch:
          type: string
        fresh:
          type: boolean
        agentLaunch:
          type: string
        assignee:
          type: string
    EntityDispatchStateDto:
      type: object
      properties:
        entityId:
          type: string
        archetype:
          type: string
        status:
          type: string
        nextPhase:
          type: string
        blocked:
          type: boolean
        dispatchable:
          type: boolean
        mode:
          $ref: "#/components/schemas/DispatchMode"
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    Response26:
      type: object
      properties:
        state:
          $ref: "#/components/schemas/EntityDispatchStateDto"
    Response27:
      type: object
      properties:
        dispatch:
          $ref: "#/components/schemas/EntityDispatchDto"
...
    TicketType:
      type: string
      enum:
      - BUG
      - IMPROVEMENT
      - MAINTENANCE
...
    TransitionedEntity:
      type: object
      properties:
        id:
          type: string
        archetype:
          $ref: "#/components/schemas/Archetype"
        projectId:
          type: string
        number:
          type: integer
          format: int64
        qualifiedId:
          type: string
        title:
          type: string
        slug:
          type: string
        slugScope:
          type: string
        description:
          type: string
        status:
          type: string
        statusBefore:
          type: string
        ticketType:
          $ref: "#/components/schemas/TicketType"
        impetus:
          type: string
        assignee:
          type: string
        createdBy:
          type: string
        supersededBy:
          type: string
        repositoryId:
          type: string
        implementedAt:
          $ref: "#/components/schemas/Instant"
        implementingAt:
          $ref: "#/components/schemas/Instant"
        dependsOn:
          type: string
        parent:
          type: string
        position:
          type: integer
          format: int32
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
        changedBy:
          type: string
        blocked:
          type: boolean
        acceptanceCriteria:
          type: array
          items:
            type: string
...
    WorkDispatchAnswer:
      type: object
      description: The dispatch a press made.
      properties:
        dispatch:
          $ref: "#/components/schemas/EntityDispatchDto"
    WorkDispatchRequest:
      type: object
      required:
      - mode
      description: Which of the two dispatch actions to press.
      properties:
        mode:
          type: string
          description: "FLOW (Dispatch: run the whole flow) or PHASE (Run the next\
            \ phase: one, then stop)"
    WorkDispatchState:
      type: object
      description: What a press would start now.
      properties:
        state:
          $ref: "#/components/schemas/EntityDispatchStateDto"
...
    WorkProgressAnswer:
      type: object
      description: "A campaign's progress, derived on the read."
      properties:
        progress:
          $ref: "#/components/schemas/CampaignProgressDto"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /projects/api/entities:
    post:
      summary: Create an entity of any archetype
      description: "Files a new epic, ticket, campaign, feature or task from one body\
        \ shape: the archetype plus the archetype's create schema (GET /projects/api/entities/archetypes/{archetype}/schemas/create).\
        \ A root names its project (id or slug), a node its parent (UUID or qualified\
        \ id): a feature's parent is an EPIC, a task's a FEATURE. The status starts\
        \ REPORTED. Answers the entity in the merged shape, with its qualified id."
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityCreate"
        required: true
      responses:
        "201":
          description: The entity as written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionedEntity"
        "400":
          description: "Every complaint about the body in one message: not an object,\
            \ no or an unknown archetype, a property the create schema does not list,\
            \ a missing required one, a value of the wrong shape, a parent of the\
            \ wrong kind, or a repository outside the project"
        "403":
          description: An agent creating in a project other than its own; nothing
            is written
        "404":
          description: "The project, the parent or the repository names nothing"
        "409":
          description: "The owning epic's status freezes its scope: a node needs the\
            \ epic REPORTED"
        "401":
          description: Not Authorized
      tags:
      - Entity Create Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
        - qits:system
...
  /projects/api/entities/{id}/dispatch:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response26"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: State
      tags:
      - Entity Dispatch Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
    post:
      summary: "Put an agent on an epic or a ticket, or start a campaign"
      description: "mode PHASE runs the one phase the entity's status starts; mode\
        \ FLOW runs that phase and the ones after it until a status starts none. The\
        \ archetype registry's phases say which phases those are, per status. On a\
        \ campaign the press is its start."
      operationId: dispatchEntity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchRequest"
        required: true
      responses:
        "200":
          description: "The dispatch made, or — for a campaign — its progress as the\
            \ start press left it"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Response27"
                - $ref: "#/components/schemas/CampaignProgressResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      tags:
      - Entity Dispatch Controller
      security:
      - SecurityScheme:
        - qits:admin
...
  /projects/api/work/{qualifiedId}/dispatch:
    get:
      summary: What a dispatch press would start now
      description: "The entity's status, the phase a press would start (or null),\
        \ whether it is blocked and dispatchable, and the mode the last press recorded.\
        \ The path names the entity by qualified id or UUID."
      operationId: getWorkDispatch
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The dispatch state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkDispatchState"
        "404":
          description: No entity with this id
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Dispatch Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
    post:
      summary: "Put an agent on an epic or a ticket, or start a campaign"
      description: mode PHASE runs the one phase the entity's status starts; mode
        FLOW runs that phase and the ones after it until a status starts none. On
        a campaign the press is its start. The path names the entity by qualified
        id (<projectSlug>-<n>) or UUID.
      operationId: dispatchWork
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkDispatchRequest"
        required: true
      responses:
        "200":
          description: "The dispatch made, or — for a campaign — its progress as the\
            \ start press left it"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/WorkDispatchAnswer"
                - $ref: "#/components/schemas/WorkProgressAnswer"
        "400":
          description: A missing or unknown mode
        "404":
          description: No entity with this id
        "409":
          description: "A status that starts no phase, a feature or a task, a blocked\
            \ entity, a project with no wrapper"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Dispatch Controller
      security:
      - SecurityScheme:
        - qits:admin
...
info:
  title: qits-projects API
  version: 2026.1006.121753